Wasser 1.1
New: App now includes a special mode for winter. By wrtiting a date in
the source you can switch to the different modes on time. Made for the
end of this season. By switching $end_time and $cur_time you can define
either the beginning of the season or the end.