Wasser 1.6.1
Recent temperatures are now queried by date and not by id, which solves
a lot of problems.
Height of the wrap is now 100%.
Error reporting is turned off.
Index detects if scraped website is offline and outputs a special error
message at the top of the application.
Temperature array slice in core is now searched backwards.
Added temperature for 14, 13, 12 and 11 degress.